backend/python/requirements.txt
backoff==2.2.1
fastapi==0.116.1
loguru==0.7.3
openai==1.98.0
pydantic==2.11.7
python-dotenv==1.1.1
slowapi==0.1.9
SQLAlchemy==2.0.42
uvicorn==0.35.0
asyncpg==0.30.0